chore: add log types package
